move fetch block api to storesdk
